Rename model from NavBarBuilder to NavBarLinks and its collaborators
