+++ src/components/NavBar/container.tsx
@@ -7,8 +7,8 @@ import { NavBar } from "./component";
 import { INavBarTranslations } from "./interface";
 import { Redirect } from "../Redirect";
 import { useCurrentUser, useLogout } from "$hooks";
-import { NavBarBuilder } from "$models/NavBarBuilder";
-import { INavBarLink } from "$models/NavBarBuilder/Interfaces";
+import { NavBarLinks } from "$models/NavBarLinks";
+import { INavBarLink } from "$models/NavBarLinks/Interfaces";
 
 export const NavBarContainer: FunctionComponent = () => {
   const history = useHistory();
@@ -24,7 +24,7 @@ export const NavBarContainer: FunctionComponent = () => {
 
   const currentUser = currentUserResponse.data.getCurrentUser;
   let links: INavBarLink[] = [];
-  if (currentUser) links = NavBarBuilder.getLinks(currentUser, translations.data);
+  if (currentUser) links = NavBarLinks.getLinks(currentUser, translations.data);
 
   const onLogOut = async () => {
     await client.clearStore();
