--- src/models/NavBarBuilder/index.ts
@@ -1,1 +1,0 @@
-export { NavBarBuilder } from "./NavBarBuilder";
