+++ src/models/NavBarLinks/CompanyNavBarLinks.ts
@@ -5,7 +5,7 @@ import { getTooltipMessage } from "./getTooltipMessage";
 import { INavBarTranslations } from "$components/NavBar/interface";
 import { INavBarLink } from "./Interfaces";
 
-export const CompanyNavBarBuilder = {
+export const CompanyNavBarLinks = {
   getLinks: (currentUser: ICurrentUser, translations: INavBarTranslations): INavBarLink[] => {
     const { jobApplications, createOffer, myOffers, myProfile } = RoutesBuilder.company;
     const { getAccessError } = Permissions;
