+++ src/models/NavBarLinks/index.ts
@@ -1,0 +1,1 @@
+export { NavBarLinks } from "./NavBarLinks";
